+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-### 3.3.70.0 (2017-03-31)
+### 3.3.71.0 (2017-03-31)
+  * AWS Resource Groups Tagging API
+    - Added support for the new Resource Groups Tagging API service. Cmdlets for the service have the noun prefix 'RGT' and can be listed using the command 'Get-AWSCmdletName -Service RGT': Add-RGTResourceTag (TagResources API), Get-RGTResource (GetResources API), Get-RGTTagKey (GetTagKeys API), Get-RGTTagValue (GetTagValues API) and Remove-RGTResourceTag (UntagResources API).
   * AWS Storage Gateway
     - Added a new cmdlet, Invoke-CacheRefresh (RefreshCache API), and extended the New-SGNFSFileShare and Update-SGNFSFileShare cmdlets with new parameters -ReadOnly and -Squash.
-    
+  * Amazon Cloud Directory
+    - Updated the Get-CDIRObjectAttribute cmdlet with two new parameters, -FacetFilter_FacetName and -FacetFilter_SchemaArn, to support the new service feature enabling filtering by facet.  
+
 ### 3.3.69.0 (2017-03-29)
   * AWS Batch
     - Added parameters to support specifying retry strategy for the Register-BATJobDefinition and Submit-BATJob cmdlets. The parameter, -RetryStrategy_Attempt, accepts a numeric value for attempts. This is the number of non successful executions before a job is considered FAILED. In addition, the JobDetail object returned by other cmdlets now has an attempts field and shows all execution attempts.